Add Log-Likelihood

Calculates and adds posterior log-likelihoods to an object of class 'bvecmodel'.

Usage

# S3 method for class 'bvecmodel'
add_posterior_loglik(object, ...)

Arguments

object

an object of class 'bvecmodel', usually, the result of a call to add_posterior_coefficients.

...

additional arguments.

Value

The object in object with posterior$loglik added, a mcmc object with one row per draw and one column per period of the training sample, holding the pointwise log-likelihood that selection_criteria uses.

See also

Examples

# Load data
data("e6")

# Create model
model <- create_bvecmodel(e6, p = 4, r = 1,
                          const = "unrestricted",
                          seasonal = "unrestricted",
                          iterations = 20, burnin = 10)
# Number of iterations and burnin should be much higher.

# Add priors
model <- add_priors(model,
                    coef = list(v_i = 1, v_i_det = 1 / 10),
                    coint = list(v_i = 0, p_tau_i = 1),
                    sigma = list(df = "k", scale = 1))

# Add initial values
model <- add_initial_values(model)

# Obtain posterior draws 
model <- add_posterior_coefficients(model)

# Add log-likelihoods
model <- add_posterior_loglik(model)