Align Observations Across Models

Restricts each model in an object of class 'modellist' to the set of observations common to all models, ensuring that comparisons are computed on the same underlying sample.

Usage

# S3 method for class 'modellist'
align_model_obs(object, ...)

Arguments

object

an object of class 'modellist'.

...

further arguments passed to or from other methods.

Value

An object of class 'modellist'.

See also

Examples

# Load data
data(e6)

# Create multiple VAR models
model_1 <- create_bvarmodel(diff(e6), p = 0:2, deterministic = "const")

# Create multiple VEC models
model_2 <- create_bvecmodel(e6, p = 1, r = 0:1, const = "unrestricted")

# Combine the models in one object
model <- combine_models(model_1, model_2)

model <- align_model_obs(model)