Combine Models

Convenience function that combines multiple models into one object.

Usage

combine_models(...)

Arguments

...

One or multiple lists that contain either model objects or posterior objects.

Value

A list of class 'modellist' that holds the models of all arguments in one flat list, in the order in which they were given.

See also

Examples

# Load data
data(e6)

# Create multiple VAR models
model_1 <- create_bvarmodel(diff(e6), p = 0:2, deterministic = "const")

# Create multiple VEC models
model_2 <- create_bvecmodel(e6, p = 1, r = 0:1, const = "unrestricted")

# Combine the models in one object
model <- combine_models(model_1, model_2)