Spillover Index

Guard method for objects of class 'bvecmodel'.

Usage

# S3 method for class 'bvecmodel'
spillover(object, ...)

Arguments

object

an object of class 'bvecmodel'.

...

additional arguments.

Value

Nothing. The method raises an error, since spillover indices of a VEC model are obtained from its VAR representation: apply vec_to_var first and then spillover to the resulting 'bvarmodel'.

See also